rename CPUTopology -> ProcessorInformation
