+++ include/firestarter/Environment/Environment.hpp
@@ -21,8 +21,8 @@
 
 #pragma once
 
-#include "firestarter/Environment/CPUTopology.hpp"
 #include "firestarter/Environment/Platform/PlatformConfig.hpp"
+#include "firestarter/Environment/ProcessorInformation.hpp"
 
 #include <cassert>
 #include <cstdint>
